requirements.txt
Babel==2.2.0
billiard==3.3.0.22
celery==3.1.20
Django==1.9.3
django-celery==3.1.17
django-debug-toolbar==1.4
django-filter==0.12.0
django-guardian==1.4.1
django-mptt==0.8.3
django-reversion==1.10.1
djangorestframework==3.3.2
psycopg2==2.6.1
python-dateutil==2.5.0
pytz==2015.7
requests==2.9.1
geoip2==2.2.0
simpleeval==0.8.7
kombu==3.0.33
maxminddb==1.2.0
amqp==1.4.9
anyjson==0.3.3
cycler==0.10.0
matplotlib==1.5.1
money==1.2.2
numpy==1.10.4
pandas==0.17.1
pycountry==1.20
pyparsing==2.1.0
scipy==0.17.0
six==1.10.0
sqlparse==0.1.18